+++ src/app/camera-capture/CosineSimilarity.js
