+++ src/app/camera-capture/PoseCombobox.js
